HW2/requirements.txt
torch==2.5.0
torchvision==0.20.0
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.1
pycocotools==2.0.8
Pillow==11.1.0
tqdm==4.65.2
opencv-python==4.10.0
ensemble-boxes==1.0.9
